refactor: OmokRules 인터페이스 추가
