+++ src/main/kotlin/omok/Main.kt
@@ -2,6 +2,7 @@ package omok
 
 import omok.controller.OmokController
 import omok.domain.board.OmokBoard
+import omok.domain.rule.OmokRules
 import omok.domain.rule.finder.DfsRenjuFinder
 import omok.domain.rule.renjuRule.RenjuRule
 import omok.domain.stone.OmokStones
@@ -10,8 +11,11 @@ import omok.view.OutputView
 
 fun main() {
     val points = OmokStones()
-    val rule = RenjuRule(DfsRenjuFinder)
-    val omokBoard = OmokBoard(points, rule)
-    val controller = OmokController(OutputView, InputView, omokBoard, rule)
+    val rules =
+        object : OmokRules {
+            override val rules = listOf(RenjuRule(DfsRenjuFinder))
+        }
+    val omokBoard = OmokBoard(points, rules)
+    val controller = OmokController(OutputView, InputView, omokBoard, rules)
     controller.startGame()
 }
