+++ src/test/kotlin/domain/fixture/FakeOmokRule.kt
@@ -1,14 +1,11 @@
 package domain.fixture
 
-import omok.domain.board.OmokBoard
 import omok.domain.rule.OmokRule
-import omok.domain.stone.Stone
+import omok.domain.rule.OmokRules
+import omok.domain.rule.finder.DfsRenjuFinder
+import omok.domain.rule.renjuRule.RenjuRule
 
-object FakeOmokRule : OmokRule {
-    override fun isProtected(
-        stone: Stone,
-        board: OmokBoard,
-    ): Boolean {
-        return true
-    }
+object FakeOmokRule : OmokRules {
+    override val rules: List<OmokRule>
+        get() = listOf(RenjuRule(DfsRenjuFinder))
 }
