+++ src/test/kotlin/domain/fixture/OmokBoardFixture.kt
@@ -1,11 +1,19 @@
 package domain.fixture
 
 import omok.domain.board.OmokBoard
+import omok.domain.rule.OmokRule
+import omok.domain.rule.OmokRules
 import omok.domain.rule.finder.DfsRenjuFinder
-import omok.domain.rule.renjuRule.RenjuRule
+import omok.domain.rule.renjuRule.SixMokRule
 import omok.domain.stone.OmokStones
 
 fun omokBoardFixture(): OmokBoard {
     val omokStones = OmokStones()
-    return OmokBoard(omokStones, RenjuRule(DfsRenjuFinder))
+    return OmokBoard(
+        omokStones,
+        object : OmokRules {
+            override val rules: List<OmokRule>
+                get() = listOf(SixMokRule(DfsRenjuFinder))
+        },
+    )
 }
